Split useStreamPlayer's pure policy and runtime monitor into lib modules
The 1,600-line hook exported ~16 pure functions solely so tests could
import them. The playback policy (stream classification, Xtream HLS
conversion, proxy URL builders, recovery-window math, live-buffer
resync/rate thresholds) now lives in lib/playback.ts, and the
self-contained runtime monitor (watchdog, stall detection, resync
timer) is a plain factory in lib/runtimeMonitor.ts operating on the
refs the hook passes in. The hook drops from 1,612 to 903 lines with
identical wiring; tests import the policy from lib/playback.
