+++ src/hooks/useStreamPlayer.ts
@@ -1,68 +1,32 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import type { ChannelResult, ContentType } from "../lib/types";
+import type { ChannelResult } from "../lib/types";
 import { normalizeCodecName, resolveResolutionLabel } from "../lib/format";
 import { logger } from "../lib/logger";
 import { toProxyUrl } from "../lib/proxyUrl";
 import { getStreamingProxyPort } from "../lib/tauri";
+import {
+  areStreamMetadataEqual,
+  classifyStream,
+  decidePlaybackRecovery,
+  formatPlaybackRecoveryMessage,
+  getMpegtsPlaybackUrls,
+  MAX_PLAYBACK_RECOVERY_ATTEMPTS,
+  readMediaErrorMessage,
+  recordPlaybackRecoveryAttempt,
+  shouldResetPlaybackRecoveryAttempts,
+  supportsNativeHlsPlayback,
+  tryConvertToXtreamHls,
+  type HlsErrorPayload,
+  type PlaybackRecoveryIssue,
+  type PlaybackStartMode,
+  type PlayerState,
+  type StreamMetadata,
+} from "../lib/playback";
+import { createRuntimeMonitor, type MpegtsPlayer } from "../lib/runtimeMonitor";
 
-type PlayerState = "idle" | "loading" | "playing" | "error";
-export type StreamType = "hls" | "mpegts" | "unknown";
-export type PlaybackStartMode = "manual" | "recovery";
-type PlaybackRecoveryIssue =
-  | "startup_failure"
-  | "media_error"
-  | "library_error"
-  | "watchdog_stall"
-  | "ended";
-
-interface PlaybackRecoveryDecisionRetry {
-  kind: "retry";
-  nextAttempt: number;
-}
-
-interface PlaybackRecoveryDecisionFail {
-  kind: "fail";
-}
-
-interface PlaybackRecoveryDecisionIgnore {
-  kind: "ignore";
-}
-
-export type PlaybackRecoveryDecision =
-  | PlaybackRecoveryDecisionRetry
-  | PlaybackRecoveryDecisionFail
-  | PlaybackRecoveryDecisionIgnore;
-
-export interface StreamMetadata {
-  width: number | null;
-  height: number | null;
-  resolution: string | null;
-  codec: string | null;
-  fps: number | null;
-  videoBitrate: string | null;
-  audioCodec: string | null;
-  audioBitrate: string | null;
-  audioOnly: boolean;
-}
-
-export function areStreamMetadataEqual(
-  left: StreamMetadata | null,
-  right: StreamMetadata | null,
-): boolean {
-  if (left === right) return true;
-  if (!left || !right) return false;
-  return (
-    left.width === right.width &&
-    left.height === right.height &&
-    left.resolution === right.resolution &&
-    left.codec === right.codec &&
-    left.fps === right.fps &&
-    left.videoBitrate === right.videoBitrate &&
-    left.audioCodec === right.audioCodec &&
-    left.audioBitrate === right.audioBitrate &&
-    left.audioOnly === right.audioOnly
-  );
-}
+// Re-exported for components (e.g. StreamPlayer) that read the recovery cap
+// alongside the hook. The implementation lives in lib/playback.
+export { MAX_PLAYBACK_RECOVERY_ATTEMPTS } from "../lib/playback";
 
 export interface UseStreamPlayerReturn {
   playerState: PlayerState;
@@ -85,133 +49,6 @@ export interface UseStreamPlayerReturn {
 
 interface UseStreamPlayerOptions {
   onPlaybackFailed?: (result: ChannelResult) => void;
-}
-
-interface PlaybackRecoveryDecisionInput {
-  issue: PlaybackRecoveryIssue;
-  recoveryTimestamps: number[];
-  now: number;
-  isPaused: boolean;
-  contentType: ContentType;
-  maxAttempts?: number;
-  windowMs?: number;
-}
-
-interface HlsErrorPayload {
-  fatal?: boolean;
-  type?: string;
-  details?: string;
-}
-
-export type HlsFatalRecoveryAction = "restart_network" | "recover_media" | "reconnect";
-
-interface MpegtsPlayer {
-  destroy(): void;
-  attachMediaElement(el: HTMLMediaElement): void;
-  load(): void;
-  on?(event: string, handler: (...args: unknown[]) => void): void;
-  off?(event: string, handler: (...args: unknown[]) => void): void;
-  mediaInfo?: {
-    videoCodec?: string;
-    audioCodec?: string;
-    width?: number;
-    height?: number;
-    hasVideo?: boolean;
-    hasAudio?: boolean;
-    fps?: number;
-    videoDataRate?: number;
-    audioDataRate?: number;
-  };
-}
-
-function stripUrlDecorations(url: string): string {
-  return url.toLowerCase().split("#", 1)[0]?.split("?", 1)[0] ?? url.toLowerCase();
-}
-
-export function classifyStream(url: string): StreamType {
-  const lower = url.toLowerCase();
-  const clean = stripUrlDecorations(url);
-  if (lower.includes(".m3u8") || clean.endsWith(".m3u8") || lower.includes("/hls/")) return "hls";
-  if (
-    clean.endsWith(".ts") ||
-    clean.endsWith(".m2ts") ||
-    clean.endsWith(".mpegts") ||
-    clean.endsWith(".m4s") ||
-    (lower.includes("/live/") && !lower.includes(".m3u8"))
-  ) {
-    return "mpegts";
-  }
-  return "unknown";
-}
-
-/**
- * Convert the common Xtream live URL forms to their HLS equivalent.
- * Providers frequently return `/live/user/pass/id.ts`; treating that as a
- * generic MPEG-TS endpoint can let a bursty response fill WebView's MSE quota.
- */
-export function tryConvertToXtreamHls(url: string): string | null {
-  let parsed: URL;
-  try {
-    parsed = new URL(url);
-  } catch {
-    return null;
-  }
-  if (parsed.protocol !== "http:" && parsed.protocol !== "https:") return null;
-
-  const segments = parsed.pathname.split("/").filter(Boolean);
-  const firstSegment = segments[0]?.toLowerCase();
-  const offset = firstSegment === "live" ? 1 : 0;
-  if (segments.length - offset !== 3) return null;
-
-  const user = segments[offset];
-  const password = segments[offset + 1];
-  const streamPart = segments[offset + 2];
-  const streamMatch = streamPart?.match(/^(\d+)(?:\.(?:ts|m2ts|mpegts))?$/i);
-  if (!user || !password || !streamMatch) return null;
-
-  parsed.pathname = `/live/${user}/${password}/${streamMatch[1]}.m3u8`;
-  return parsed.toString();
-}
-
-function toStreamingProxyUrl(
-  url: string,
-  port: number,
-  reconnect: boolean,
-  remux: boolean,
-): string {
-  const reconnectParam = reconnect ? "&reconnect=1" : "";
-  const remuxParam = remux ? "&remux=1" : "";
-  return `http://127.0.0.1:${port}/stream?url=${encodeURIComponent(url)}${reconnectParam}${remuxParam}`;
-}
-
-export function getMpegtsPlaybackUrls(
-  url: string,
-  proxyPort: number,
-  isLive: boolean,
-): string[] {
-  if (proxyPort <= 0) {
-    return [url];
-  }
-
-  const proxyUrl = toStreamingProxyUrl(url, proxyPort, isLive, false);
-  return isLive
-    ? [toStreamingProxyUrl(url, proxyPort, true, true), proxyUrl]
-    : [proxyUrl];
-}
-
-export function shouldSuspendPlaybackWatchdog(
-  visibilityState: DocumentVisibilityState,
-): boolean {
-  return visibilityState === "hidden";
-}
-
-export function supportsNativeHlsPlayback(
-  mediaElement: Pick<HTMLMediaElement, "canPlayType">,
-): boolean {
-  return [
-    "application/vnd.apple.mpegurl",
-    "application/x-mpegurl",
-  ].some((mimeType) => mediaElement.canPlayType(mimeType) !== "");
 }
 
 function readStoredVolume(): number {
@@ -243,211 +80,9 @@ function createVideoElement(): HTMLVideoElement {
   return el;
 }
 
-function readMediaErrorMessage(mediaErr: MediaError | null): string | null {
-  if (!mediaErr) return null;
-  const codeMap: Record<number, string> = {
-    1: "Playback aborted",
-    2: "Network error",
-    3: "Decode error",
-    4: "Format not supported",
-  };
-  return codeMap[mediaErr.code] ?? mediaErr.message ?? "Unknown media error";
-}
-
-export const MAX_PLAYBACK_RECOVERY_ATTEMPTS = 5;
-export const PLAYBACK_RECOVERY_WINDOW_MS = 2 * 60_000;
 const LOADING_TIMEOUT_MS = 15_000;
 const NATIVE_HLS_TIMEOUT_MS = 4_000;
 const PLAYBACK_RECOVERY_DELAY_MS = 900;
-const PLAYBACK_STALL_GRACE_MS = 15_000;
-const PLAYBACK_NO_PROGRESS_STALL_MS = 20_000;
-const PLAYBACK_WATCHDOG_POLL_MS = 1_000;
-const MIN_PROGRESS_DELTA_SECS = 0.05;
-const HLS_BUFFER_STALLED_ERROR = "bufferStalledError";
-const HLS_FATAL_RECOVERY_MAX_ATTEMPTS = 2;
-const HLS_FATAL_RECOVERY_WINDOW_MS = 30_000;
-const LIVE_RESYNC_MIN_SEEK_SECS = 1;
-const LIVE_RESYNC_MAX_SEEK_SECS = 3;
-const LIVE_RESYNC_MAX_JUMP_SECS = 30;
-const LIVE_RESYNC_COOLDOWN_MS = 3_000;
-const LIVE_RESYNC_WAIT_CONFIRM_MS = 750;
-const LIVE_RESYNC_SILENT_STALL_MS = 3_000;
-const LIVE_BUFFER_SLOWDOWN_THRESHOLD_SECS = 12;
-const LIVE_BUFFER_RECOVERED_THRESHOLD_SECS = 20;
-const LIVE_BUFFER_RECOVERY_RATE = 0.97;
-const LIVE_BUFFER_CATCHUP_THRESHOLD_SECS = 30;
-const LIVE_BUFFER_CATCHUP_RATE = 1.08;
-const LIVE_BUFFER_MAX_LATENCY_SECS = 60;
-const LIVE_BUFFER_TARGET_LATENCY_SECS = 20;
-
-export interface BufferedTimeRange {
-  start: number;
-  end: number;
-}
-
-export function chooseLiveBufferPlaybackRate(
-  currentRate: number,
-  bufferedAhead: number,
-): number {
-  if (!Number.isFinite(bufferedAhead)) return 1;
-  if (currentRate < 1) {
-    return bufferedAhead < LIVE_BUFFER_RECOVERED_THRESHOLD_SECS
-      ? LIVE_BUFFER_RECOVERY_RATE
-      : 1;
-  }
-  if (currentRate > 1) {
-    return bufferedAhead > LIVE_BUFFER_RECOVERED_THRESHOLD_SECS
-      ? LIVE_BUFFER_CATCHUP_RATE
-      : 1;
-  }
-  if (bufferedAhead > LIVE_BUFFER_CATCHUP_THRESHOLD_SECS) {
-    return LIVE_BUFFER_CATCHUP_RATE;
-  }
-  return bufferedAhead < LIVE_BUFFER_SLOWDOWN_THRESHOLD_SECS
-    ? LIVE_BUFFER_RECOVERY_RATE
-    : 1;
-}
-
-function bufferedSecondsAhead(currentTime: number, ranges: BufferedTimeRange[]): number {
-  const currentRange = ranges.find(
-    (range) => currentTime >= range.start && currentTime < range.end,
-  );
-  return currentRange ? Math.max(0, currentRange.end - currentTime) : 0;
-}
-
-export function findLiveLatencyCatchUpTarget(
-  currentTime: number,
-  ranges: BufferedTimeRange[],
-  maxLatency = LIVE_BUFFER_MAX_LATENCY_SECS,
-  targetLatency = LIVE_BUFFER_TARGET_LATENCY_SECS,
-): number | null {
-  if (!Number.isFinite(currentTime)) return null;
-  const currentRange = ranges.find(
-    (range) => currentTime >= range.start && currentTime < range.end,
-  );
-  if (!currentRange || currentRange.end - currentTime <= maxLatency) {
-    return null;
-  }
-  const target = Math.max(
-    currentRange.start + MIN_PROGRESS_DELTA_SECS,
-    currentRange.end - targetLatency,
-  );
-  return target > currentTime + MIN_PROGRESS_DELTA_SECS ? target : null;
-}
-
-/**
- * Choose a safe point already buffered by MSE when a live stream stops at a
- * timestamp/keyframe discontinuity. Some IPTV MPEG-TS sources keep delivering
- * data while leaving the playhead on an undecodable boundary; rebuilding the
- * whole player loses far more video than moving to the next buffered keyframe.
- */
-export function findLiveBufferResyncTarget(
-  currentTime: number,
-  ranges: BufferedTimeRange[],
-  maxJump = LIVE_RESYNC_MAX_JUMP_SECS,
-): number | null {
-  if (!Number.isFinite(currentTime) || ranges.length === 0) return null;
-
-  for (const range of ranges) {
-    if (!Number.isFinite(range.start) || !Number.isFinite(range.end) || range.end <= range.start) {
-      continue;
-    }
-
-    if (range.start > currentTime + MIN_PROGRESS_DELTA_SECS) {
-      const gap = range.start - currentTime;
-      if (gap > maxJump) return null;
-      return Math.min(range.end - 0.01, range.start + MIN_PROGRESS_DELTA_SECS);
-    }
-
-    if (
-      currentTime >= range.start - MIN_PROGRESS_DELTA_SECS &&
-      currentTime < range.end - MIN_PROGRESS_DELTA_SECS
-    ) {
-      const availableAdvance = range.end - currentTime - MIN_PROGRESS_DELTA_SECS;
-      const advance = Math.min(
-        LIVE_RESYNC_MAX_SEEK_SECS,
-        maxJump,
-        Math.max(LIVE_RESYNC_MIN_SEEK_SECS, availableAdvance),
-      );
-      const target = Math.min(range.end - MIN_PROGRESS_DELTA_SECS, currentTime + advance);
-      return target > currentTime + MIN_PROGRESS_DELTA_SECS ? target : null;
-    }
-  }
-
-  return null;
-}
-
-export function getHlsFatalRecoveryAction(type?: string): HlsFatalRecoveryAction {
-  if (type === "networkError") return "restart_network";
-  if (type === "mediaError") return "recover_media";
-  return "reconnect";
-}
-
-export function getNextPlaybackRecoveryAttempt(
-  recoveryTimestamps: number[],
-  now: number,
-  maxAttempts = MAX_PLAYBACK_RECOVERY_ATTEMPTS,
-  windowMs = PLAYBACK_RECOVERY_WINDOW_MS,
-): number | null {
-  const nextAttempt =
-    prunePlaybackRecoveryHistory(recoveryTimestamps, now, windowMs).length + 1;
-  return nextAttempt <= maxAttempts ? nextAttempt : null;
-}
-
-export function prunePlaybackRecoveryHistory(
-  recoveryTimestamps: number[],
-  now: number,
-  windowMs = PLAYBACK_RECOVERY_WINDOW_MS,
-): number[] {
-  return recoveryTimestamps.filter((timestamp) => now - timestamp < windowMs);
-}
-
-export function recordPlaybackRecoveryAttempt(
-  recoveryTimestamps: number[],
-  now: number,
-  windowMs = PLAYBACK_RECOVERY_WINDOW_MS,
-): number[] {
-  return [...prunePlaybackRecoveryHistory(recoveryTimestamps, now, windowMs), now];
-}
-
-export function shouldResetPlaybackRecoveryAttempts(
-  startMode: PlaybackStartMode,
-  previousChannelIndex: number | null,
-  nextChannelIndex: number,
-): boolean {
-  return startMode === "manual" || previousChannelIndex !== nextChannelIndex;
-}
-
-export function formatPlaybackRecoveryMessage(
-  attempt: number,
-  maxAttempts = MAX_PLAYBACK_RECOVERY_ATTEMPTS,
-): string {
-  return `Stream interrupted. Reconnecting (${attempt}/${maxAttempts})...`;
-}
-
-export function decidePlaybackRecovery(
-  input: PlaybackRecoveryDecisionInput,
-): PlaybackRecoveryDecision {
-  if (input.isPaused) {
-    return { kind: "ignore" };
-  }
-  if (input.issue === "ended" && input.contentType !== "live") {
-    return { kind: "ignore" };
-  }
-  const nextAttempt = getNextPlaybackRecoveryAttempt(
-    input.recoveryTimestamps,
-    input.now,
-    input.maxAttempts,
-    input.windowMs,
-  );
-  if (nextAttempt === null) {
-    return { kind: "fail" };
-  }
-  return {
-    kind: "retry",
-    nextAttempt,
-  };
-}
 
 type StartPlaybackAttempt = (
   result: ChannelResult,
@@ -795,363 +430,19 @@ export function useStreamPlayer(options?: UseStreamPlayerOptions): UseStreamPlay
   const setupRuntimeMonitor = useCallback(
     (result: ChannelResult, sessionId: number) => {
       cleanupRuntimeMonitor();
-
-      let closed = false;
-      const monitor = {
-        lastProgressAt: performance.now(),
-        lastCurrentTime: videoElement.currentTime,
-        stallStartedAt: null as number | null,
-        lastResyncAt: Number.NEGATIVE_INFINITY,
-      };
-      let resyncTimer: ReturnType<typeof setTimeout> | null = null;
-      let hlsFatalRecoveryTimestamps: number[] = [];
-
-      const handlers: Array<{ target: EventTarget; event: string; handler: EventListener }> = [];
-      const addHandler = (target: EventTarget, event: string, handler: EventListener) => {
-        target.addEventListener(event, handler);
-        handlers.push({ target, event, handler });
-      };
-
-      const markProgress = () => {
-        if (closed) return;
-        const now = performance.now();
-        const currentTime = videoElement.currentTime;
-        if (currentTime > monitor.lastCurrentTime + MIN_PROGRESS_DELTA_SECS) {
-          monitor.lastCurrentTime = currentTime;
-          monitor.stallStartedAt = null;
-        }
-        monitor.lastProgressAt = now;
-      };
-
-      const markPotentialStall = () => {
-        if (closed || isPausedRef.current || videoElement.paused) {
-          return;
-        }
-        if (monitor.stallStartedAt === null) {
-          monitor.stallStartedAt = performance.now();
-        }
-      };
-
-      const tryLiveBufferResync = () => {
-        if (
-          closed ||
-          !mpegtsPlayerRef.current ||
-          isPausedRef.current ||
-          videoElement.paused ||
-          performance.now() - monitor.lastResyncAt < LIVE_RESYNC_COOLDOWN_MS
-        ) {
-          return false;
-        }
-
-        const ranges: BufferedTimeRange[] = [];
-        for (let index = 0; index < videoElement.buffered.length; index += 1) {
-          ranges.push({
-            start: videoElement.buffered.start(index),
-            end: videoElement.buffered.end(index),
-          });
-        }
-        const from = videoElement.currentTime;
-        const target = findLiveBufferResyncTarget(from, ranges);
-        if (target === null || target <= from + MIN_PROGRESS_DELTA_SECS) {
-          return false;
-        }
-
-        monitor.lastResyncAt = performance.now();
-        monitor.lastCurrentTime = target;
-        monitor.lastProgressAt = monitor.lastResyncAt;
-        monitor.stallStartedAt = null;
-        logger.warn(
-          `[Player] Skipping ${(target - from).toFixed(2)}s buffered timestamp gap`,
-        );
-        videoElement.currentTime = target;
-        void videoElement.play().catch(() => {});
-        return true;
-      };
-
-      const adjustLiveBufferPlaybackRate = () => {
-        if (!mpegtsPlayerRef.current) return;
-        const ranges: BufferedTimeRange[] = [];
-        for (let index = 0; index < videoElement.buffered.length; index += 1) {
-          ranges.push({
-            start: videoElement.buffered.start(index),
-            end: videoElement.buffered.end(index),
-          });
-        }
-        const bufferedAhead = bufferedSecondsAhead(videoElement.currentTime, ranges);
-        const nextRate = chooseLiveBufferPlaybackRate(videoElement.playbackRate, bufferedAhead);
-        if (Math.abs(videoElement.playbackRate - nextRate) < 0.001) return;
-        videoElement.playbackRate = nextRate;
-        logger.info(
-          nextRate < 1
-            ? `[Player] Slowing live playback to ${nextRate.toFixed(2)}x to rebuild buffer`
-            : nextRate > 1
-              ? `[Player] Increasing live playback to ${nextRate.toFixed(2)}x to reduce latency`
-            : "[Player] Restored live playback to 1.00x",
-        );
-      };
-
-      const trimExcessiveLiveLatency = () => {
-        if (!mpegtsPlayerRef.current || videoElement.buffered.length === 0) return false;
-        const ranges: BufferedTimeRange[] = [];
-        for (let index = 0; index < videoElement.buffered.length; index += 1) {
-          ranges.push({
-            start: videoElement.buffered.start(index),
-            end: videoElement.buffered.end(index),
-          });
-        }
-        const from = videoElement.currentTime;
-        const target = findLiveLatencyCatchUpTarget(from, ranges);
-        if (target === null) return false;
-
-        const now = performance.now();
-        monitor.lastResyncAt = now;
-        monitor.lastCurrentTime = target;
-        monitor.lastProgressAt = now;
-        monitor.stallStartedAt = null;
-        logger.warn(
-          `[Player] Trimming ${(target - from).toFixed(1)}s accumulated live latency`,
-        );
-        videoElement.currentTime = target;
-        void videoElement.play().catch(() => {});
-        return true;
-      };
-
-      const clearResyncTimer = () => {
-        if (!resyncTimer) return;
-        clearTimeout(resyncTimer);
-        resyncTimer = null;
-      };
-
-      const scheduleLiveBufferResync = () => {
-        markPotentialStall();
-        if (resyncTimer) return;
-        resyncTimer = setTimeout(() => {
-          resyncTimer = null;
-          tryLiveBufferResync();
-        }, LIVE_RESYNC_WAIT_CONFIRM_MS);
-      };
-
-      const triggerRuntimeIssue = (
-        issue: PlaybackRecoveryIssue,
-        reason: string,
-      ) => {
-        if (closed || playbackSessionIdRef.current !== sessionId || !hasStartedPlayingRef.current) {
-          return;
-        }
-        closed = true;
-        attemptRecoveryOrFail(result, sessionId, issue, reason);
-      };
-
-      addHandler(videoElement, "playing", () => {
-        clearResyncTimer();
-        hlsFatalRecoveryTimestamps = [];
-        monitor.lastProgressAt = performance.now();
-        monitor.lastCurrentTime = videoElement.currentTime;
-        monitor.stallStartedAt = null;
+      runtimeMonitorCleanupRef.current = createRuntimeMonitor({
+        result,
+        sessionId,
+        videoElement,
+        hlsInstanceRef,
+        mpegtsPlayerRef,
+        isPausedRef,
+        playerStateRef,
+        playbackSessionIdRef,
+        hasStartedPlayingRef,
+        onRuntimeIssue: (issue, reason) =>
+          attemptRecoveryOrFail(result, sessionId, issue, reason),
       });
-      addHandler(videoElement, "timeupdate", () => {
-        markProgress();
-      });
-      addHandler(videoElement, "canplay", () => {
-        clearResyncTimer();
-        hlsFatalRecoveryTimestamps = [];
-        markProgress();
-        monitor.stallStartedAt = null;
-      });
-      addHandler(videoElement, "waiting", () => {
-        scheduleLiveBufferResync();
-      });
-      addHandler(videoElement, "stalled", () => {
-        scheduleLiveBufferResync();
-      });
-      addHandler(videoElement, "pause", () => {
-        if (closed || isPausedRef.current || videoElement.ended) return;
-        logger.warn("[Player] Resuming an unexpected media pause");
-        window.setTimeout(() => {
-          if (
-            !closed &&
-            !isPausedRef.current &&
-            videoElement.paused &&
-            !videoElement.ended
-          ) {
-            void videoElement.play().catch(() => {});
-          }
-        }, 100);
-      });
-      addHandler(videoElement, "error", () => {
-        const reason = readMediaErrorMessage(videoElement.error) ?? "Media error during playback";
-        triggerRuntimeIssue("media_error", reason);
-      });
-      addHandler(videoElement, "ended", () => {
-        if (result.content_type !== "live") {
-          return;
-        }
-        triggerRuntimeIssue("ended", "Live stream ended unexpectedly");
-      });
-
-      const libCleanups: Array<() => void> = [];
-
-      const hls = hlsInstanceRef.current;
-      if (hls) {
-        const hlsEvents = hls as unknown as {
-          on(event: string, handler: (...args: unknown[]) => void): void;
-          off(event: string, handler: (...args: unknown[]) => void): void;
-        };
-        const onHlsError = (...args: unknown[]) => {
-          const data = args[1] as HlsErrorPayload | undefined;
-          if (!data) {
-            return;
-          }
-          if (data.details === HLS_BUFFER_STALLED_ERROR) {
-            markPotentialStall();
-          }
-          if (data.fatal) {
-            const detail = data.details ?? "fatal hls.js error";
-            const type = data.type ?? "hls.js";
-            const recoveryAction = getHlsFatalRecoveryAction(data.type);
-            if (recoveryAction !== "reconnect") {
-              const now = performance.now();
-              const attempt = getNextPlaybackRecoveryAttempt(
-                hlsFatalRecoveryTimestamps,
-                now,
-                HLS_FATAL_RECOVERY_MAX_ATTEMPTS,
-                HLS_FATAL_RECOVERY_WINDOW_MS,
-              );
-              if (attempt === null) {
-                triggerRuntimeIssue(
-                  "library_error",
-                  `${type}: repeated fatal recovery failed (${detail})`,
-                );
-                return;
-              }
-              hlsFatalRecoveryTimestamps = recordPlaybackRecoveryAttempt(
-                hlsFatalRecoveryTimestamps,
-                now,
-                HLS_FATAL_RECOVERY_WINDOW_MS,
-              );
-            }
-            if (recoveryAction === "restart_network") {
-              logger.warn("[Player] Restarting hls.js network loading after", detail);
-              hls.startLoad(-1);
-              markPotentialStall();
-              return;
-            }
-            if (recoveryAction === "recover_media") {
-              logger.warn("[Player] Recovering hls.js media pipeline after", detail);
-              hls.recoverMediaError();
-              markPotentialStall();
-              return;
-            }
-            triggerRuntimeIssue("library_error", `${type}: ${detail}`);
-          }
-        };
-        const onHlsStallResolved = () => {
-          hlsFatalRecoveryTimestamps = [];
-          markProgress();
-          monitor.stallStartedAt = null;
-        };
-        hlsEvents.on("hlsError", onHlsError);
-        hlsEvents.on("hlsStallResolved", onHlsStallResolved);
-        libCleanups.push(() => {
-          try {
-            hlsEvents.off("hlsError", onHlsError);
-            hlsEvents.off("hlsStallResolved", onHlsStallResolved);
-          } catch {}
-        });
-      }
-
-      const mpegtsPlayer = mpegtsPlayerRef.current;
-      if (mpegtsPlayer?.on && mpegtsPlayer.off) {
-        const onMpegtsError = (
-          errorType?: unknown,
-          errorDetail?: unknown,
-          info?: unknown,
-        ) => {
-          const segments = [errorType, errorDetail, info]
-            .filter((value): value is string => typeof value === "string" && value.length > 0);
-          const detail = segments.join(": ") || "mpegts.js runtime error";
-          triggerRuntimeIssue("library_error", detail);
-        };
-        mpegtsPlayer.on("error", onMpegtsError);
-        libCleanups.push(() => {
-          try {
-            mpegtsPlayer.off?.("error", onMpegtsError);
-          } catch {}
-        });
-      }
-
-      const watchdog = setInterval(() => {
-        if (closed || playbackSessionIdRef.current !== sessionId) {
-          return;
-        }
-        if (shouldSuspendPlaybackWatchdog(document.visibilityState)) {
-          monitor.lastProgressAt = performance.now();
-          monitor.stallStartedAt = null;
-          return;
-        }
-        if (playerStateRef.current !== "playing") {
-          return;
-        }
-        if (
-          isPausedRef.current ||
-          videoElement.paused ||
-          videoElement.seeking ||
-          videoElement.ended
-        ) {
-          return;
-        }
-
-        const now = performance.now();
-        if (trimExcessiveLiveLatency()) {
-          return;
-        }
-        adjustLiveBufferPlaybackRate();
-        const currentTime = videoElement.currentTime;
-        if (currentTime > monitor.lastCurrentTime + MIN_PROGRESS_DELTA_SECS) {
-          monitor.lastCurrentTime = currentTime;
-          monitor.lastProgressAt = now;
-          monitor.stallStartedAt = null;
-          return;
-        }
-
-        if (monitor.stallStartedAt !== null) {
-          if (
-            now - monitor.stallStartedAt >= LIVE_RESYNC_WAIT_CONFIRM_MS &&
-            tryLiveBufferResync()
-          ) {
-            return;
-          }
-          if (now - monitor.stallStartedAt >= PLAYBACK_STALL_GRACE_MS) {
-            triggerRuntimeIssue("watchdog_stall", "Stream stalled during playback");
-          }
-          return;
-        }
-
-        const noProgressDuration = now - monitor.lastProgressAt;
-        if (
-          noProgressDuration >= LIVE_RESYNC_SILENT_STALL_MS &&
-          tryLiveBufferResync()
-        ) {
-          return;
-        }
-
-        if (noProgressDuration >= PLAYBACK_NO_PROGRESS_STALL_MS) {
-          triggerRuntimeIssue("watchdog_stall", "Playback stopped progressing");
-        }
-      }, PLAYBACK_WATCHDOG_POLL_MS);
-
-      runtimeMonitorCleanupRef.current = () => {
-        closed = true;
-        clearInterval(watchdog);
-        clearResyncTimer();
-        videoElement.playbackRate = 1;
-        for (const h of handlers) {
-          h.target.removeEventListener(h.event, h.handler);
-        }
-        for (const fn of libCleanups) {
-          fn();
-        }
-      };
     },
     [attemptRecoveryOrFail, cleanupRuntimeMonitor, videoElement],
   );
